+++ src/testIntegration/java/tests/pages/IT8_Pages_Lock.java
@@ -9,10 +9,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import testkit.WithEmptyTestPage;
 
-public class IT40_Pages_Lock extends WithEmptyTestPage {
+public class IT8_Pages_Lock extends WithEmptyTestPage {
 
   @Test
-  @DisplayName("IT-40: Pages - Lock and unlock a page")
+  @DisplayName("IT-8: Pages - Lock and unlock a page")
   public void testLockAndUnlockPage() {
     Page locked =
         getNotionClient()